requirements.txt
# packages in environment at /home/jw/.applications/miniconda/envs/pyqtdev:
#
# Name                    Version                   Build  Channel
altgraph                  0.15                       py_0    conda-forge
astroid                   1.6.3                    py36_0  
autopep8                  1.3.5                     <pip>
ca-certificates           2018.4.16                     0    conda-forge
certifi                   2018.4.16                py36_0    conda-forge
dbus                      1.13.2               h714fa37_1  
expat                     2.2.5                he0dffb1_0  
fontconfig                2.12.6               h49f89f6_0  
freetype                  2.8                  hab7d2ae_1  
future                    0.16.0                    <pip>
glib                      2.56.1               h000015b_0  
gst-plugins-base          1.14.0               hbbd80ab_1  
gstreamer                 1.14.0               hb453b48_1  
icu                       58.2                 h9c2bf20_1  
isort                     4.3.4                    py36_0  
jpeg                      9b                   h024ee3a_2  
lazy-object-proxy         1.3.1            py36h10fcdad_0  
libedit                   3.1                  heed3624_0  
libffi                    3.2.1                hd88cf55_4  
libgcc-ng                 7.2.0                hdf63c60_3  
libpng                    1.6.34               hb9fc6fc_0  
libstdcxx-ng              7.2.0                hdf63c60_3  
libxcb                    1.13                 h1bed415_1  
libxml2                   2.9.8                hf84eae3_0  
macholib                  1.8                        py_0    conda-forge
mccabe                    0.6.1            py36h5ad9710_1  
ncurses                   6.0                  h9df7e31_2  
openssl                   1.0.2o                        0    conda-forge
pcre                      8.42                 h439df22_0  
pefile                    2017.11.5                 <pip>
pip                       10.0.1                   py36_0  
pycodestyle               2.4.0                     <pip>
pycrypto                  2.6.1                    py36_1    conda-forge
pyinstaller               3.3.1                    py36_0    conda-forge
pylint                    1.8.4                    py36_0  
pyqt                      5.9.2            py36h751905a_0  
pyqtdeploy                2.1                       <pip>
PySide2                   5.9.0a1.dev1525349078           <pip>
python                    3.6.5                hc3d631a_2  
qt                        5.9.5                h7e424d6_0  
readline                  7.0                  ha6073c6_4  
setuptools                39.1.0                   py36_0  
sip                       4.19.8           py36hf484d3e_0  
six                       1.11.0           py36h372c433_1  
sqlite                    3.23.1               he433501_0  
tk                        8.6.7                hc745277_3  
wheel                     0.31.0                   py36_0  
wrapt                     1.10.11          py36h28b7045_0  
xz                        5.2.3                h5e939de_4  
zlib                      1.2.11               ha838bed_2  
